fix: handle large volume of images
